+++ ChatBox/AppFiles/Assets.xcassets/AccentColor.colorset/Contents.json
