+++ ChatBox/AppFiles/Assets.xcassets/AppIcon.appiconset/Contents.json
