+++ ChatBox/AppFiles/Assets.xcassets/AppIcon.appiconset/mac256.png
