+++ ChatBox/AppFiles/Assets.xcassets/AppIcon.appiconset/mac32.png
