+++ ChatBox/AppFiles/Assets.xcassets/AppIcon.appiconset/mac512.png
