+++ ChatBox/AppFiles/Assets.xcassets/Contents.json
