+++ ChatBox/AppFiles/Assets.xcassets/LaunchScreen/BackgroundImage.imageset/Contents.json
