+++ ChatBox/AppFiles/Assets.xcassets/LaunchScreen/Contents.json
