--- ChatBox/AuthScene/Model/AuthResponse.swift
@@ -1,12 +1,0 @@
-//
-//  AuthResponse.swift
-//  ChatBox
-//
-//  Created by Alexander Chervoncev on 7/2/2023.
-//
-
-import Foundation
-
-enum AuthResponse {
-    case success, errorAccountNotVerified, errorLogin, error
-}
