--- ChatBox/RegisterScene/Service/RegisterResponse.swift
@@ -1,15 +1,0 @@
-//
-//  RegisterResponse.swift
-//  ChatBox
-//
-//  Created by Alexander Chervoncev on 6/2/2023.
-//
-
-import Foundation
-
-enum RegisterResponse {
-    case success, emailAlreadyInUse, nicknameAlreadyInUse, error, unknownError
-}
-
-
-
